refactor: :recycle: updated some inconsistencies, splitted user and book methods giving each their respective folder
