+++ src/Storytel.ts
@@ -1,5 +1,5 @@
-import { login } from "./methods/login.js";
-import { revalidateAccount } from "./methods/revalidateAccount.js";
+import { login } from "./methods/user/login.js";
+import { revalidateAccount } from "./methods/user/revalidateAccount.js";
 import { User } from "./User.js";
 
 import type { SingleSignToken } from "./types/types.js";
