+++ src/methods/user/getAccountInfo.ts
@@ -1,5 +1,5 @@
-import { Account } from "../types/account";
-import type { SingleSignToken } from "../types/types";
+import { Account } from "../../types/account";
+import type { SingleSignToken } from "../../types/types";
 
 const URL = "https://www.storytel.com/api/getPersonAccountInfo.action?token={TOKEN}";
 
@@ -7,13 +7,13 @@ export async function getAccountInfo(token: SingleSignToken): Promise<Account> {
 	const formattedURL = URL.replace("{TOKEN}", token);
 
 	const response = await fetch(formattedURL);
-	const JSON = await response.json();
+	const json = await response.json();
 
-	if (response.status !== 200 || !JSON?.personInfo) {
-		throw new Error(JSON.message);
+	if (response.status !== 200 || !json?.personInfo) {
+		throw new Error(json.message);
 	}
 
-	const accountInfo = JSON.personInfo as Account;
+	const accountInfo = json.personInfo as Account;
 
 	return accountInfo;
 }
