+++ src/methods/user/login.ts
@@ -1,5 +1,5 @@
-import { encryptPassword } from "../utils/utils.js";
-import type { LoginResponse } from "../types/types.js";
+import { encryptPassword } from "../../utils/utils.js";
+import type { LoginResponse } from "../../types/types.js";
 
 const URL = "https://www.storytel.com/api/login.action?m=1&uid={UID}&pwd={PASSWORD}";
 
@@ -10,11 +10,11 @@ export async function login(email: string, password: string) {
 		.replace("{PASSWORD}", encrypted);
 
 	const response = await fetch(formattedURL);
-	const JSON = await response.json() as LoginResponse;
+	const json = await response.json() as LoginResponse;
 
 	if (response.status !== 200) {
-		throw new Error(JSON.message);
+		throw new Error(json.message);
 	}
 
-	return JSON;
+	return json;
 }
